Allow grouped parameters in Parameters table def
This feature allows parameter entries in a parameter table to be sent to
multiple receiving components, so long as the entries share the same
type. Parameters can now be listed like:

parameters:
   - Comp1.Par1
   - [Comp1.Par2, Comp2.Par1]

where the second entry in the table is now served to two destination
components. Each component will receive the same value and must have the
same parameter type.