Add on_change mode to Product Packetizer generator
This feature allows a packet to only be emitted if one of the data
product timestamps in that packet is newer than the last time that
packet was emitted. On-change packets are still created at the
configured period, but are dropped immediately if all the data products
have timestamps older than the timestamp of the last packet sent.